requirements.txt
requests==2.24.0
python-telegram-bot==12.8
googletrans==3.0.0